Add restart mechanism in examples/ChemistryTests/HomogeneousBox/run.sh

+++ examples/ChemistryTests/HomogeneousBox/run.sh
@@ -10,6 +10,7 @@ rho=${rho:=0.1} # Gas density
 run_name=${run_name:=""}
 with_cosmo=${with_cosmo:=0}
 with_initial_metals=${with_initial_metals:=0} # Do we want some initial pollution?
+with_restart=${with_restart:=0}
 
 # Choose the right parameters
 if [ "$with_initial_metals" -eq 1 ];
@@ -50,22 +51,30 @@ then
 	swift_run_parameters="$swift_run_parameters --cosmology"
 fi
 
+# Add restart runtime parameter to the executable
+if [ "$with_restart" -eq 1 ]; then
+	echo "Running with cosmology"
+	swift_run_parameters="$swift_run_parameters --restart"
+fi
+
 echo "Parameter file: $parameter_file"
 
 # Remove the ICs
-if [ -e ICs_homogeneous_box.hdf5 ]
-then
-    echo "------------------------------"
-	echo "Removing ICs..."
-    rm ICs_homogeneous_box.hdf5
-fi
+if [ "$with_restart" -eq 0 ]; then
+	if [ -e ICs_homogeneous_box.hdf5 ]
+	then
+		echo "------------------------------"
+		echo "Removing ICs..."
+		rm ICs_homogeneous_box.hdf5
+	fi
 
-#Create the ICs if they do not exist
-if [ ! -e ICs_homogeneous_box.hdf5 ]
-then
-    echo "Generating initial conditions to run the example..."
-    python3 makeIC.py --level $level -o ICs_homogeneous_box.hdf5 \
-	    --lJ $jeans_length --rho $rho --epsilon $radius --metal_mass_fraction $metal_mass_fraction
+	#Create the ICs if they do not exist
+	if [ ! -e ICs_homogeneous_box.hdf5 ]
+	then
+		echo "Generating initial conditions to run the example..."
+		python3 makeIC.py --level $level -o ICs_homogeneous_box.hdf5 \
+			--lJ $jeans_length --rho $rho --epsilon $radius --metal_mass_fraction $metal_mass_fraction
+	fi
 fi
 
 
@@ -86,20 +95,23 @@ fi
 
 # Create output directory
 DIR=snap #First test of units conversion
-if [ -d "$DIR" ];
-then
-    echo "$DIR directory exists. Its content will be removed."
-    rm -r $DIR
-else
-    echo "$DIR directory does not exists. It will be created."
-    mkdir $DIR
+
+if [ "$with_restart" -eq 0 ]; then
+	if [ -d "$DIR" ];
+	then
+		echo "$DIR directory exists. Its content will be removed."
+		rm -r $DIR
+	else
+		echo "$DIR directory does not exists. It will be created."
+		mkdir $DIR
+	fi
 fi
 
 printf "Running simulation..."
 
 
 ../../../swift --hydro --stars --self-gravity --feedback \
-	       --cooling --limiter $swift_run_parameters  \
+	       --cooling --limiter $swift_run_parameters --pin  \
 	       --threads=$n_threads $parameter_file 2>&1 | tee output.log
 
 #Do some data analysis to show what's in this box
